Move the scene creation from Game to its own class Scene
