+++ src/main/java/nl/tudelft/pixelperfect/Game.java
@@ -14,6 +14,7 @@ import com.jme3.network.Server;
 import com.jme3.network.serializing.Serializer;
 import com.jme3.scene.Geometry;
 import com.jme3.scene.shape.Box;
+import nl.tudelft.pixelperfect.Scene;
 
 import nl.tudelft.pixelperfect.client.ConnectListener;
 import nl.tudelft.pixelperfect.client.HelloMessage;
@@ -35,6 +36,7 @@ public class Game extends SimpleApplication {
   private Spaceship spaceship;
   private EventScheduler scheduler;
   private Server server;
+  private Scene scene;
 
   /**
    * Main method bootstrapping the process by constructing this class and initializing a
@@ -53,10 +55,11 @@ public class Game extends SimpleApplication {
    */
   @Override
   public void simpleInitApp() {
+    scene = new Scene(this);
+    scene.createMap();
     // increase movement speed
     flyCam.setMoveSpeed(50);
-    createMap();
-    // Spatial map = assetManager.loadModel("assets/Models/cockpit1/cockpit1.j3o");
+
     try {
       server = Network.createServer(6143);
       Serializer.registerClass(HelloMessage.class);
@@ -71,93 +74,6 @@ public class Game extends SimpleApplication {
     scheduler = new EventScheduler(0.5);
 
     scheduler.subscribe(spaceship.getLog());
-  }
-
-  /**
-   * Method that contains all objects for the scene.
-   */
-  public void createMap() {
-    drawDashboard();
-    drawFloor();
-    drawWalls();
-    drawTimer();
-  }
-
-  /**
-   * Render the dashboard of the scene.
-   */
-  public void drawDashboard() {
-    Box dashboard = new Box(4, 1f, 1);
-    Geometry geom = new Geometry("Box", dashboard);
-    Material mat = new Material(assetManager, "Common/MatDefs/Misc/Unshaded.j3md");
-    mat.setColor("Color", ColorRGBA.Blue);
-    geom.setMaterial(mat);
-    geom.setLocalTranslation(new Vector3f(0, 0, -3));
-    rootNode.attachChild(geom);
-  }
-
-  /**
-   * Render the floor of the scene.
-   */
-  public void drawFloor() {
-    // green floor
-    Box floor = new Box(4, 0.01f, 4);
-    Geometry g2 = new Geometry("Floor", floor);
-    Material mat2 = new Material(assetManager, "Common/MatDefs/Misc/Unshaded.j3md");
-    mat2.setColor("Color", ColorRGBA.Green);
-    g2.setMaterial(mat2);
-    g2.setLocalTranslation(0, -1, 0);
-    rootNode.attachChild(g2);
-  }
-
-  /**
-   * Render the walls of the scene.
-   */
-  public void drawWalls() {
-    // walls
-    Box wallLeft = new Box(0.01f, 4, 4);
-    Geometry l1 = new Geometry("leftwall", wallLeft);
-    Material mat3 = new Material(assetManager, "Common/MatDefs/Misc/Unshaded.j3md");
-    mat3.setColor("Color", ColorRGBA.Orange);
-    l1.setMaterial(mat3);
-    l1.setLocalTranslation(new Vector3f(-4, 3, 0));
-    rootNode.attachChild(l1);
-
-    Box wallRight = new Box(0.01f, 4, 4);
-    Geometry l2 = new Geometry("rightwall", wallRight);
-    l2.setMaterial(mat3);
-    l2.setLocalTranslation(new Vector3f(4, 3, 0));
-    rootNode.attachChild(l2);
-
-    Box wallFront = new Box(4, 4, 0.01f);
-    Geometry l3 = new Geometry("frontwall", wallFront);
-    Material mat4 = new Material(assetManager, "Common/MatDefs/Misc/Unshaded.j3md");
-    mat4.setColor("Color", ColorRGBA.Yellow);
-    l3.setMaterial(mat4);
-    l3.setLocalTranslation(new Vector3f(0, 3, -4));
-    rootNode.attachChild(l3);
-
-    Box wallBack = new Box(4, 4, 0.01f);
-    Geometry l4 = new Geometry("backwall", wallBack);
-    l4.setMaterial(mat4);
-    l4.setLocalTranslation(new Vector3f(0, 3, 4));
-    rootNode.attachChild(l4);
-  }
-
-  /**
-   * Render placeholder for timer that will be displayed.
-   */
-  public void drawTimer() {
-    guiNode.detachAllChildren();
-    guiFont = assetManager.loadFont("Interface/Fonts/Default.fnt");
-    BitmapText timer = new BitmapText(guiFont, false);
-    // timer.setSize(1);
-    timer.setText("mm:ss");
-    timer.setLocalTranslation(2.5f, 5, 3.9f);
-    timer.setLocalScale(0.1f);
-    timer.setLocalRotation(
-        new Quaternion().fromAngleAxis(180 * FastMath.DEG_TO_RAD, new Vector3f(0, 1, 0)));
-    rootNode.attachChild(timer);
   }
 
   /**
